refactor(resample): separate common code
